+++ lib/recording_studio_accessible/authorization_class_methods.rb
@@ -68,9 +68,17 @@ module RecordingStudioAccessible
       return RecordingStudio::Recording.none unless root_access_recordings.exists?
 
       relation = RecordingStudio::Recording.unscoped
-                                           .where(id: root_access_recordings.select(:root_recording_id))
+                                           .where(id: effective_root_recording_ids(root_access_recordings))
                                            .distinct
       exclude_shared_roots(relation)
+    end
+
+    def effective_root_recording_ids(root_access_recordings)
+      return root_access_recordings.select(:root_recording_id) unless DependentAccess.column_available?
+
+      root_access_recordings.includes(:recordable).filter_map do |access_recording|
+        access_recording.root_recording_id if DependentAccess.effective?(access_recording)
+      end
     end
 
     def exclude_shared_roots(relation)